refactor(NodesTable): decompose into `NodeVersion` component
